Refactor: extract log_processor.go
