+++ handlers/analytics.go
@@ -1,13 +1,13 @@
 package handlers
 
 import (
-	"bytes"
 	"encoding/json"
 	"errors"
 	"fmt"
 	"github.com/golang/glog"
 	"github.com/julienschmidt/httprouter"
 	cerrors "github.com/livepeer/catalyst-api/errors"
+	"github.com/livepeer/catalyst-api/handlers/analytics"
 	mistapiconnector "github.com/livepeer/catalyst-api/mapic"
 	"github.com/livepeer/go-api-client"
 	"github.com/mileusna/useragent"
@@ -16,16 +16,12 @@ import (
 	"io"
 	"net/http"
 	"strconv"
-	"strings"
 	"sync"
-	"time"
 )
 
 const (
 	GeoHashPrecision        = 3
 	MaxConcurrentProcessing = 5000
-	SendMetricsInterval     = 10 * time.Second
-	SendMetricsTimeout      = 60 * time.Second
 )
 
 type AnalyticsHandler struct {
@@ -67,18 +63,7 @@ type AnalyticsExternalData struct {
 	UserID string
 }
 
-type AnalyticsData struct {
-	sessionID  string
-	playbackID string
-	browser    string
-	deviceType string
-	country    string
-	userID     string
-}
-
 type AnalyticsHandlersCollection struct {
-	schema *gojsonschema.Schema
-
 	streamCache mistapiconnector.IStreamCache
 	lapi        *api.Client
 
@@ -91,7 +76,6 @@ type AnalyticsHandlersCollection struct {
 
 func NewAnalyticsHandlersCollection(streamCache mistapiconnector.IStreamCache, lapi *api.Client, metricsURL string, host string) AnalyticsHandlersCollection {
 	return AnalyticsHandlersCollection{
-		schema:      inputSchemasCompiled["AnalyticsLog"],
 		streamCache: streamCache,
 		lapi:        lapi,
 		cache:       make(map[string]AnalyticsExternalData),
@@ -101,11 +85,14 @@ func NewAnalyticsHandlersCollection(streamCache mistapiconnector.IStreamCache, l
 }
 
 func (c *AnalyticsHandlersCollection) Log() httprouter.Handle {
-	dataCh := make(chan AnalyticsData, MaxConcurrentProcessing)
-	c.startLogProcessor(dataCh)
+	schema := inputSchemasCompiled["AnalyticsLog"]
+
+	dataCh := make(chan analytics.AnalyticsData, MaxConcurrentProcessing)
+	lp := analytics.NewLogProcessor(c.metricsURL, c.host)
+	lp.Start(dataCh)
 
 	return func(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
-		log, err := parseAnalyticsLog(r, c.schema)
+		log, err := parseAnalyticsLog(r, schema)
 		if log == nil {
 			cerrors.WriteHTTPBadRequest(w, "Invalid request payload", err)
 			return
@@ -247,15 +234,15 @@ func (c *AnalyticsHandlersCollection) cacheExtData(playbackID string, extData An
 	return extData, nil
 }
 
-func toAnalyticsData(log *AnalyticsLog, geo *AnalyticsGeo, extData AnalyticsExternalData) AnalyticsData {
+func toAnalyticsData(log *AnalyticsLog, geo *AnalyticsGeo, extData AnalyticsExternalData) analytics.AnalyticsData {
 	ua := useragent.Parse(log.UserAgent)
-	return AnalyticsData{
-		sessionID:  log.SessionID,
-		playbackID: log.PlaybackID,
-		browser:    ua.Name,
-		deviceType: deviceTypeOf(ua),
-		country:    geo.Country,
-		userID:     extData.UserID,
+	return analytics.AnalyticsData{
+		SessionID:  log.SessionID,
+		PlaybackID: log.PlaybackID,
+		Browser:    ua.Name,
+		DeviceType: deviceTypeOf(ua),
+		Country:    geo.Country,
+		UserID:     extData.UserID,
 	}
 }
 
@@ -268,120 +255,4 @@ func deviceTypeOf(ua useragent.UserAgent) string {
 		return "desktop"
 	}
 	return ""
-}
-
-type LogProcessor struct {
-	logs    map[labelsKey]map[string]metricValue
-	promURL string
-	host    string
-}
-
-type metricValue struct {
-}
-
-type labelsKey struct {
-	playbackID string
-	browser    string
-	deviceType string
-	country    string
-	userID     string
-}
-
-func NewLogProcessor(promURL string, host string) LogProcessor {
-	return LogProcessor{
-		logs:    make(map[labelsKey]map[string]metricValue),
-		promURL: promURL,
-		host:    host,
-	}
-}
-
-func (c *AnalyticsHandlersCollection) startLogProcessor(ch chan AnalyticsData) {
-	t := time.NewTicker(SendMetricsInterval)
-	lp := NewLogProcessor(c.metricsURL, c.host)
-
-	go func() {
-		for {
-			select {
-			case d := <-ch:
-				lp.processLog(d)
-			case <-t.C:
-				lp.sendMetrics()
-			}
-		}
-	}()
-}
-
-func (p *LogProcessor) processLog(d AnalyticsData) {
-	var k = labelsKey{
-		playbackID: d.playbackID,
-		browser:    d.browser,
-		deviceType: d.deviceType,
-		country:    d.country,
-		userID:     d.userID,
-	}
-
-	bySessionID, ok := p.logs[k]
-	if !ok {
-		p.logs[k] = make(map[string]metricValue)
-		bySessionID = p.logs[k]
-	}
-	bySessionID[d.sessionID] = metricValue{}
-}
-
-func (p *LogProcessor) sendMetrics() {
-	if len(p.logs) > 0 {
-		glog.Info("sending analytics logs")
-	} else {
-		glog.V(6).Info("no analytics logs, skip sending")
-		return
-	}
-
-	// convert values in the Prometheus format
-	var metrics strings.Builder
-	now := time.Now().UnixMilli()
-	for k, v := range p.logs {
-		metrics.WriteString(p.toMetric(k, v, now))
-	}
-
-	// send data
-	err := p.sendMetricsString(metrics.String())
-	if err != nil {
-		glog.Errorf("failed to send analytics logs, err=%w", err)
-	}
-
-	// clear map
-	p.logs = make(map[labelsKey]map[string]metricValue)
-}
-
-func (p *LogProcessor) toMetric(k labelsKey, v map[string]metricValue, nowMs int64) string {
-	return fmt.Sprintln(fmt.Sprintf(`viewcount{host="%s",user_id="%s",playback_id="%s",device_type="%s",browser="%s",country="%s"} %d %d`,
-		p.host,
-		k.userID,
-		k.playbackID,
-		k.deviceType,
-		k.browser,
-		k.country,
-		len(v),
-		nowMs,
-	))
-}
-
-func (p *LogProcessor) sendMetricsString(metrics string) error {
-	client := &http.Client{Timeout: SendMetricsTimeout}
-	req, err := http.NewRequest("POST", p.promURL, bytes.NewBuffer([]byte(metrics)))
-	if err != nil {
-		return err
-	}
-	req.Header.Set("Content-Type", "text/plain")
-
-	resp, err := client.Do(req)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK && resp.StatusCode != http.StatusNoContent {
-		return fmt.Errorf("non-OK status code: %d", resp.StatusCode)
-	}
-	return nil
 }
