+++ handlers/analytics_test.go
@@ -4,7 +4,6 @@ import (
 	"github.com/julienschmidt/httprouter"
 	"github.com/livepeer/go-api-client"
 	"github.com/stretchr/testify/require"
-	"io"
 	"net/http"
 	"net/http/httptest"
 	"strings"
@@ -234,119 +233,4 @@ func TestEnrichExtData(t *testing.T) {
 	require.NoError(err)
 	require.Equal(userID, res.UserID)
 	require.Equal(1, mockMapicCache.callCount)
-}
-
-func TestLogProcessor(t *testing.T) {
-	require := require.New(t)
-
-	tests := []struct {
-		name     string
-		data     []AnalyticsData
-		expected []string
-	}{
-		{
-			name: "Multiple same logs results in 1 metric",
-			data: []AnalyticsData{
-				{
-					sessionID:  "session-1",
-					playbackID: "playback-1",
-					browser:    "chrome",
-					deviceType: "mobile",
-					country:    "Poland",
-					userID:     "user-1",
-				},
-				{
-					sessionID:  "session-1",
-					playbackID: "playback-1",
-					browser:    "chrome",
-					deviceType: "mobile",
-					country:    "Poland",
-					userID:     "user-1",
-				},
-			},
-			expected: []string{
-				`viewcount{user_id="user-1",playback_id="playback-1",device_type="mobile",browser="chrome",country="Poland"} 1`,
-			},
-		},
-		{
-			name: "Multiple same logs with different session IDs results in 1 metric",
-			data: []AnalyticsData{
-				{
-					sessionID:  "session-1",
-					playbackID: "playback-1",
-					browser:    "chrome",
-					deviceType: "mobile",
-					country:    "Poland",
-					userID:     "user-1",
-				},
-				{
-					sessionID:  "session-2",
-					playbackID: "playback-1",
-					browser:    "chrome",
-					deviceType: "mobile",
-					country:    "Poland",
-					userID:     "user-1",
-				},
-			},
-			expected: []string{
-				`viewcount{user_id="user-1",playback_id="playback-1",device_type="mobile",browser="chrome",country="Poland"} 2`,
-			},
-		},
-		{
-			name: "Different logs result in separate metrics",
-			data: []AnalyticsData{
-				{
-					sessionID:  "session-1",
-					playbackID: "playback-1",
-					browser:    "chrome",
-					deviceType: "mobile",
-					country:    "Poland",
-					userID:     "user-1",
-				},
-				{
-					sessionID:  "session-2",
-					playbackID: "playback-1",
-					browser:    "firefox",
-					deviceType: "mobile",
-					country:    "Poland",
-					userID:     "user-1",
-				},
-			},
-			expected: []string{
-				`viewcount{user_id="user-1",playback_id="playback-1",device_type="mobile",browser="chrome",country="Poland"} 1`,
-				`viewcount{user_id="user-1",playback_id="playback-1",device_type="mobile",browser="firefox",country="Poland"} 1`,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			// given
-			var recordedRequest string
-			promMockServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-				body, err := io.ReadAll(r.Body)
-				require.NoError(err)
-				recordedRequest = string(body)
-			}))
-			defer promMockServer.Close()
-
-			lp := NewLogProcessor(promMockServer.URL)
-
-			// when
-			for _, d := range tt.data {
-				lp.processLog(d)
-			}
-			lp.sendMetrics()
-
-			// then
-			recordedLines := strings.Split(recordedRequest, "\n")
-			if recordedLines[len(recordedLines)-1] == "" {
-				recordedLines = recordedLines[:len(recordedLines)-1]
-			}
-
-			require.Equal(len(tt.expected), len(recordedLines))
-			for _, exp := range tt.expected {
-				require.Contains(recordedRequest, exp)
-			}
-		})
-	}
 }
